requirements.txt
bokeh==2.4.2
numpy==1.21.2
pandas==1.3.4
ptvsd==4.3.2
python-dotenv==0.19.2
requests==2.26.0
